*** empty log message ***

+++ KlayGE/Demos/media/Electro/Electro.fx
@@ -26,12 +26,11 @@ VS_OUTPUT ElectroVS(VS_INPUT input)
 
 sampler3D electroSampler;
 
-float4 ElectroPS(float3 texCoord0	: TEXCOORD0,
-
+half4 ElectroPS(float3 texCoord0	: TEXCOORD0,
 					uniform sampler3D electroMap) : COLOR
 {
-	float turb = tex3D(electroMap, texCoord0) * 2 - 1;
-	float4 grow = (1 - pow(turb, 0.2)) * float4(1.70, 1.48, 1.78, 1);
+	half turb = tex3D(electroMap, texCoord0) * 2 - 1;
+	half4 grow = (1 - pow(turb, 0.2)) * half4(1.70, 1.48, 1.78, 1);
 	return pow(grow, 4);
 }
